🏗️ refactor(controller): ranger les 8 contrôleurs racine dans Api/ ou Web/
src/Controller/ ne contient plus aucun fichier à sa racine — poursuite
du rangement déjà commencé (Api/, Web/) : chaque contrôleur suit sa
nature réelle plutôt que son ancienneté dans le projet.

- Api/ (7 déplacés) : AlbumAddMediaController, AlbumRemoveMediaController,
  FileDownloadController, FileUploadController, MediaThumbnailController,
  RefreshTokenController, ResetPasswordController — tous exposent des
  routes /api/v1/* ou /api/* (ResetPasswordController garde aussi ses
  2 routes web /reset-password*, non scindé : rangement de fichiers,
  pas refactor de logique)
- Web/ (1 déplacé) : FolderBrowserController — route /web/folders

Références corrigées : import dans FileOutput.php (controller
API Platform), chemin de classe dans config/routes/folder_browser.yaml.
Renommages via git mv, historique préservé. Aucune régression (616 tests).
