+++ src/Controller/Api/FileUploadController.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Controller;
+namespace App\Controller\Api;
 
 use App\ApiResource\FileOutput;
 use App\Message\MediaProcessMessage;
