+++ src/Controller/Api/RefreshTokenController.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Controller;
+namespace App\Controller\Api;
 
 use App\Entity\RefreshToken;
 use App\Repository\RefreshTokenRepository;
